+++ docs/references/changelog.md
@@ -16,6 +16,8 @@ versioning][semver].
 - `kele-list` buffer now displays the Last Updated time
 - Resource kind completion is now grouped by API group. For example, Deployments, DaemonSets, and
   ReplicaSets are all grouped under `apps` in the completion buffer.
+- Resource name completion is now grouped by namespace for functions like `kele-get` or
+  `kele-delete` that operate on a single resource.
 
 ### Fixed